pyproject.toml
[tool.poetry]
name = "learn-project"
version = "0.1.0"
description = ""
authors = ["Dmitry"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.1"
aiohttp = "^3.11.7"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
black = "^24.10.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.2.1"
fastapi-cache2 = "^0.2.2"
celery = "^5.4.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
